requirements.txt
psycopg2-binary==2.9.9
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
python-dateutil==2.8.2
pytz==2021.1
elastic-apm==6.8.1
Werkzeug==2.0.1
elastic-apm==6.8.1
jaeger-client==4.6.0
flask_opentracing==1.1.0
opentelemetry-sdk==1.11.1
opentelemetry-exporter-otlp==1.11.1
protobuf<=3.20.x
blinker==1.4
